betterposter.cls: add % so latex doesn't insert \par on newline
From stackexchange: https://tex.stackexchange.com/
questions/7453/what-is-the-use-of-percent-signs-at-
the-end-of-lines

The short answer is what others have said, % starts a
comment that goes to the end of the line. The normal effect
is that it doesn't insert the space (or a \par) from the
newline. The longer answer is that as TeX parses its input,
it reads the input file line by line. It strips off tailing
whitespace (including any carriage return and newline) and
then appends a character corresponding to the number in the
\endlinechar register (provided it isn't -1 or greater than
255). By default, the value is 13 which corresponds to an
ASCII carriage return. (TeX denotes this by ^^M.) You can
change this register to be any value. Unless the category
codes have been changed (e.g., by the \obeylines macro), ^^M
has category code 5, or end of line. When TeX reaches an end
of line character, it discards the rest of the line and does
one of three things: nothing, insert a space token, or
insert a \par token, depending on what state TeX was in (S,
M, or N, respectively). So what does this have to do with %?
Well, since the comment character causes TeX to ignore the
rest of the input line, the end of line character that got
appended gets ignored too. This can frequently be important
when playing around with category codes of ^^M (again, using
\obeylines or similar).